Changing the return type of adding an entry and fixing access
Older methods were public for no good reason. This has been fixed as to not compromise the security of the database. Moreover, the return type of adding an entry has been changed for the same reason. Tests have been edited to suit the new structure, branch coverage is still at 100%.